Re-drop dnsmasq in zesty, in favor of systemd-resolved
But don't Recommend libnss-resolve, because the local resolver + libnss_dns.so
should suffice for all uses and the resolved nss module is unnecessary added
complexity.

We need to specify "dns=systemd-resolved" as for the time being our
/etc/resolv.conf points to resolvconf's generated file instead of
systemd-resolved's, so the auto-detection does not work. Do that in a separate
conf.d file in /usr to avoid unnecessary conffile prompts in the future. The
user can still override it in /etc.